+++ noxfile.py
@@ -9,17 +9,22 @@ import nox
 GCC_VERSIONS = ["gcc-5", "gcc-6", "gcc-8", "gcc-9", "clang-10"]
 GCC_VERSION2USE = os.path.split(os.environ.get("CC", "gcc-5"))[1]
 DEFAULT_TEST_DIRECTORIES = ["doc", "gcovr"]
+DEFAULT_LINT_DIRECTORIES = ["admin"] + DEFAULT_TEST_DIRECTORIES
 BLACK_CONFORM_FILES = [
+    "admin/add_copyright.py",
     "noxfile.py",
     "gcovr/decision_analysis.py",
     "gcovr/gcov.py",
     "gcovr/gcov_parser.py",
     "gcovr/timestamps.py",
+    "gcovr/tests/test_gcov_parser.py",
     "gcovr/writer/json.py",
+    "gcovr/tests/test_gcovr.py",
 ]
+BLACK_PINNED_VERSION = "black==22.1.0"
 
 nox.options.sessions = ["qa"]
-nox.options.reuse_existing_virtualenvs = True
+
 
 def set_environment(session: nox.Session, cc: str, check: bool = True) -> None:
     if check and (shutil.which(cc) is None):
@@ -50,18 +55,18 @@ def qa(session: nox.Session) -> None:
 @nox.session
 def lint(session: nox.Session) -> None:
     """Run the lint (flake8 and black)."""
-    session.install("flake8")
+    session.install("flake8", "flake8-print")
     # Black installs under Pypy but doesn't necessarily run (cf psf/black#2559).
     if platform.python_implementation() == "CPython":
-        session.install("black")
+        session.install(BLACK_PINNED_VERSION)
     if session.posargs:
         args = session.posargs
     else:
-        args = DEFAULT_TEST_DIRECTORIES
+        args = DEFAULT_LINT_DIRECTORIES
     session.run("flake8", *args)
 
     if platform.python_implementation() == "CPython":
-        if session.posargs:
+        if not session.posargs:
             session.run(
                 "python", "-m", "black", "--diff", "--check", *BLACK_CONFORM_FILES
             )
@@ -75,14 +80,12 @@ def lint(session: nox.Session) -> None:
 @nox.session
 def black(session: nox.Session) -> None:
     """Run black, a code formatter and format checker."""
-    session.install("black")
+    session.install(BLACK_PINNED_VERSION)
     if session.posargs:
         session.run("python", "-m", "black", *session.posargs)
     else:
-        session.run(
-            "python", "-m", "black", "--diff", "--check", *BLACK_CONFORM_FILES
-        )
-        session.run("python", "-m", "black", "--diff", *DEFAULT_TEST_DIRECTORIES)
+        session.run("python", "-m", "black", "--diff", "--check", *BLACK_CONFORM_FILES)
+        session.run("python", "-m", "black", "--diff", *DEFAULT_LINT_DIRECTORIES)
 
 
 @nox.session
@@ -123,6 +126,8 @@ def tests_compiler(session: nox.Session, version: str) -> None:
         "cmake",
         "yaxmldiff",
     )
+    if platform.system() == "Windows":
+        session.install("pywin32")
     coverage_args = []
     if os.environ.get("USE_COVERAGE") == "true":
         session.install("pytest-cov")
@@ -169,7 +174,7 @@ def build_wheel(session: nox.Session) -> None:
     session.notify("check_wheel")
 
 
-@nox.session(reuse_venv=False)
+@nox.session
 def check_wheel(session: nox.Session) -> None:
     """Check the wheel, should not be used directly."""
     session.install("wheel", "twine")
@@ -188,6 +193,7 @@ def upload_wheel(session: nox.Session) -> None:
 
 def docker_container_os(session: nox.Session) -> str:
     return f"ubuntu:{'18.04' if session.env['CC'] in ['gcc-5', 'gcc-6'] else '20.04'}"
+
 
 def docker_container_id(session: nox.Session, version: str) -> str:
     """Get the docker container ID."""
@@ -262,7 +268,9 @@ def docker_qa_run_compiler(session: nox.Session, version: str) -> None:
         session.env["NOX_POSARGS"] = shlex.join(session.posargs)
     else:
         # Code for join taken from Python 3.9
-        session.env["NOX_POSARGS"] = ' '.join(shlex.quote(arg) for arg in session.posargs)
+        session.env["NOX_POSARGS"] = " ".join(
+            shlex.quote(arg) for arg in session.posargs
+        )
     session.run(
         "docker",
         "run",